requirements.txt
arxiv
biopython
google-generativeai
requests
python-dotenv
